bah picky

+++ zone/client.cpp
@@ -11983,7 +11983,7 @@ void Client::SetSpellDuration(
 	}
 }
 
-void Client::SetBaseClass(uint8 class_id) {
+void Client::SetBaseClass(uint32 class_id) {
 	// Save Current Data
 	Save();
 	// Query Multiclass data for new class
